refactor: Cambiar el nombre de AuthHandler a AuthService y la interfaz a IAuthService
